# babson-scheduler/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Babson Library Student Worker Scheduler",
    "description": "\n    Constraint-aware scheduling system for Babson Library student workers.\n\n    Uses Google OR-Tools CP-SAT for optimization-based schedule generation.\n    Designed for IBM Orchestrate integration via OpenAPI custom extension.\n    ",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://your-deployment-host.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/api/v1/diagnostics/preflight": {
      "post": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Run Preflight Check",
        "operationId": "run_preflight_check_api_v1_diagnostics_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Run Preflight Check Api V1 Diagnostics Preflight Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/preflight/{week_start_date}": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Get Preflight Snapshot",
        "operationId": "get_preflight_snapshot_api_v1_diagnostics_preflight__week_start_date__get",
        "parameters": [
          {
            "name": "week_start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Week Start Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnostics/config": {
      "get": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_v1_diagnostics_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerConfigOut"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Diagnostics"
        ],
        "summary": "Update Config",
        "operationId": "update_config_api_v1_diagnostics_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchedulerConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulerConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnosticFinding": {
        "properties": {
          "check_type": {
            "type": "string",
            "title": "Check Type"
          },
          "severity": {
            "type": "string",
            "enum": [
              "hard",
              "warning",
              "info"
            ],
            "title": "Severity"
          },
          "shift_instance_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Instance Id"
          },
          "shift_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Date"
          },
          "shift_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Time"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "student_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Name"
          },
          "available_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Count"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "recommended_action": {
            "type": "string",
            "title": "Recommended Action"
          }
        },
        "type": "object",
        "required": [
          "check_type",
          "severity",
          "description",
          "recommended_action"
        ],
        "title": "DiagnosticFinding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PreflightReportOut": {
        "properties": {
          "week_start_date": {
            "type": "string",
            "title": "Week Start Date"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "is_feasible": {
            "type": "boolean",
            "title": "Is Feasible",
            "description": "False if any hard diagnostic findings exist"
          },
          "hard_findings_count": {
            "type": "integer",
            "title": "Hard Findings Count"
          },
          "warning_findings_count": {
            "type": "integer",
            "title": "Warning Findings Count"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticFinding"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "week_start_date",
          "snapshot_id",
          "created_at",
          "is_feasible",
          "hard_findings_count",
          "warning_findings_count",
          "findings"
        ],
        "title": "PreflightReportOut"
      },
      "PreflightRequest": {
        "properties": {
          "week_start_date": {
            "type": "string",
            "title": "Week Start Date"
          },
          "is_exam_period": {
            "type": "boolean",
            "title": "Is Exam Period",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "week_start_date"
        ],
        "title": "PreflightRequest"
      },
...
      "SchedulerConfigOut": {
        "properties": {
          "min_hours_default": {
            "type": "integer",
            "title": "Min Hours Default"
          },
          "max_hours_default": {
            "type": "integer",
            "title": "Max Hours Default"
          },
          "max_consecutive_hours": {
            "type": "integer",
            "title": "Max Consecutive Hours"
          },
          "shift_block_minutes": {
            "type": "integer",
            "title": "Shift Block Minutes"
          },
          "stagger_overlap_minutes": {
            "type": "integer",
            "title": "Stagger Overlap Minutes"
          },
          "hard_shift_windows": {
            "items": {},
            "type": "array",
            "title": "Hard Shift Windows"
          },
          "exam_min_hours": {
            "type": "integer",
            "title": "Exam Min Hours"
          },
          "exam_max_hours": {
            "type": "integer",
            "title": "Exam Max Hours"
          },
          "solver_time_limit_seconds": {
            "type": "integer",
            "title": "Solver Time Limit Seconds"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "min_hours_default",
          "max_hours_default",
          "max_consecutive_hours",
          "shift_block_minutes",
          "stagger_overlap_minutes",
          "hard_shift_windows",
          "exam_min_hours",
          "exam_max_hours",
          "solver_time_limit_seconds",
          "updated_at"
        ],
        "title": "SchedulerConfigOut"
      },
      "SchedulerConfigUpdate": {
        "properties": {
          "min_hours_default": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Hours Default"
          },
          "max_hours_default": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hours Default"
          },
          "max_consecutive_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Consecutive Hours"
          },
          "shift_block_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 360.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Shift Block Minutes"
          },
          "stagger_overlap_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Stagger Overlap Minutes"
          },
          "exam_min_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Min Hours"
          },
          "exam_max_hours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 40.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Max Hours"
          },
          "solver_time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 300.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Solver Time Limit Seconds"
          }
        },
        "type": "object",
        "title": "SchedulerConfigUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}